add file pageview
