--- lib/views/component/banner.dart
@@ -1,11 +1,0 @@
-class Banner {
-  final int id;
-  final String thumnail;
-
-  Banner(this.id, this.thumnail);
-}
-List<Banner> bannerList = [
-   Banner(1, "assets/images/Group 120.png"),
-   Banner(2, "assets/images/Group 120.png"),
-   Banner(2, "assets/images/Group 120.png"),
-];
