+++ lib/views/components_app/background_page.dart
@@ -1,19 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../constants.dart';
+import '../dependents_app/constants.dart';
 
 class BackgroundPage extends StatelessWidget {
   const BackgroundPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    ScreenUtil.init(context, designSize: const Size(375.0, 812.0));
     return Container(
-      // padding: EdgeInsets.only(bottom: 406.h, top: 0.0,),
       height: 812.0.h,
       width: 375.0.w,
       decoration:  BoxDecoration(
-        // color: Color(0XFFFFD223),
         gradient: kLineargradientBackgroundColors,
       ),
       child: Container(
