+++ lib/views/screens/home_page/home_page.dart
@@ -1,7 +1,7 @@
 
-import 'package:accum_points_app/views/component/background_page.dart';
-import 'package:accum_points_app/views/component/banner.dart';
-import 'package:accum_points_app/views/screens/home_page/widgets/brands/brands.dart';
+import 'package:accum_points_app/views/components/background_page.dart';
+import 'package:accum_points_app/views/screens/home_page/widgets/app_banner/banner_item.dart';
+import 'package:accum_points_app/views/screens/home_page/widgets/partners/brands.dart';
 import 'package:accum_points_app/views/screens/home_page/widgets/welcome.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -11,46 +11,21 @@ class HomePage extends StatefulWidget {
   State<HomePage> createState() => _HomePageState();
 }
 class _HomePageState extends State<HomePage> {
-
-  final controller = PageController(initialPage: 1);
-  
   @override
   Widget build(BuildContext context) {
-    ScreenUtil.init(context, designSize: const Size(375.0, 812.0));
     return Scaffold(
       body: Stack(
         children: [
     const  BackgroundPage(),
     const  Welcome(),
-       //  SizedBox(height: 16.h,),
            Container(
             padding: EdgeInsets.only(top: 130.h,),
             alignment: AlignmentDirectional.topStart,
               child: SingleChildScrollView(
                 child: Column(
                   children: [
-                    Container(
-                      width: 375.w,
-                      height: 154.h,
-                      child: PageView.builder(
-                        itemCount: bannerList.length,
-                        controller: controller,
-                        itemBuilder: ((context, index) {
-                          return Container(
-                            margin:  EdgeInsets.symmetric(horizontal: 16.w),
-                             decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(10.w),
-                               image:  DecorationImage(
-                                image: AssetImage(bannerList[index].thumnail),
-                                fit: BoxFit.cover,
-                          )
-                           ),
-                         );
-                        }
-                        ),
-                        ),
-                    ),
-                    SizedBox(height: 16.h),
+            const   BannerItem(),
+                    SizedBox(height: 8.h),
               const Brands(),
                   ],
                 ),
